Made the world infinite to the right of the camera
When the ship moves beyond a certain x direction, its x position
ceases to change and the other objects positions' change to give
the illusion that the ship is moving to the right. An array stores
all non player objects and updates them all if the ship reaches the
set certain threshold.

I've also done some cleaning up of the formatting to restrict all
lines to a maximum of 79 characters.