Added an array of hit circles to each object and debug graphics
To the creation function of each object I've added a hitCircles array. The array contains objects with x, y and r variables. These are used to define the hit circles each object has.

Also added some code to draw the circles to the screen each update step if the render_body_debug_info parameter is set.