Implemented circular hit box collision detection
When the AABB hit boxes overlap, a further test is done to check the distances between all the circles in the hitCircles array of one object with all circles in the hitCircles array of the other. If the distance is less than the circles radii combined, then a collision is initiated.